src/test/java/ng/com/smartcity/recipeApp/converters/UnitOfMeasureToUnitOfMeasureCommandTest.java
package ng.com.smartcity.recipeApp.converters;

import ng.com.smartcity.recipeApp.commands.UnitOfMeasureCommand;
import ng.com.smartcity.recipeApp.domain.UnitOfMeasure;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class UnitOfMeasureToUnitOfMeasureCommandTest {

    private final static Long ID = 1L;
    private final static String DESCRIPTION = "description";

    UnitOfMeasureToUnitOfMeasureCommand uomConverter;

    @BeforeEach
    void setUp() {
        uomConverter = new UnitOfMeasureToUnitOfMeasureCommand();
    }

    @Test
    void testNullObject() {
        assertNull(uomConverter.convert(null));
    }

    @Test
    void convert() {
        //given
        UnitOfMeasure uom = new UnitOfMeasure();
        uom.setId(ID);
        uom.setDescription(DESCRIPTION);

        //then
        UnitOfMeasureCommand command = uomConverter.convert(uom);

        assertNotNull(command);
        assertEquals(command.getId(), ID);
    }
}